refactor(reservations/availability): extract pure slot-rules module (#2628)
Pure rule functions (estimateDuration, filterSuitableTables, parseOperatingHours,
checkTableConflict, checkPacingForSlot, selectBestTable) moved to slot-rules.ts
with no Prisma import. availability.ts is now IO-only; re-exports pure fns for
backward compat. Callers updated to import directly from slot-rules.ts.

- New slot-rules.ts: 6 pure functions, 0 DB deps
- New slot-rules.test.ts: 48 tests (all GREEN)
- availability.ts: IO orchestrator only; scheduleForDay + findSuitableTables
  dedup resolved
- availability.test.ts: pure-fn describe blocks removed; IO blocks retained
- assert-bookable, confirm-hold, hold, reservation: import from slot-rules.js

Closes #2628
